requirements.txt
pandas
docopt
python-dotenv
bs4
requests
sklearn
jupyter
seaborn
progressbar2
rope
numpy
spacy
https://github.com/allanwright/media-classifier-core/archive/v1.1.3.tar.gz
azure-storage-queue
azure-storage-blob